feat: Enable image cropping and upload functionality
This commit enables the image cropping and upload functionality in the `api.py` file. The code previously had a TODO comment to continue testing the image cropping before enabling it, but now it has been uncommented. The cropped image is now being uploaded to the storage bucket. This change improves the functionality of the application by allowing users to crop and upload images of food items.