refactor: Update shelf_life_days field in FoodItemDetails to be optional
The shelf_life_days field in the FoodItemDetails class has been updated to be optional. This change allows for cases where the shelf life of a food item is unknown or not applicable. By setting the default value to None, users can now leave the shelf life days field empty when adding a new food item. This improves the flexibility and usability of the application.